aftab/AI Loan Document/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
python-dotenv==1.0.0
langchain==0.0.350
pypdf==3.17.1
python-magic==0.4.27
boto3==1.29.3
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
openai==1.3.3
anthropic==0.7.4
cohere==4.32 